# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Easy SKS API
  description: Open-source flashcard study app for the Sportkuestenschifferschein
    (SKS)
  version: 1.0.0
paths:
...
  /cards/{card_id}:
    get:
      tags:
      - Cards
      summary: Get Card
      operationId: get_card_cards__card_id__get
      parameters:
      - name: card_id
        in: path
        required: true
        schema:
          type: string
          title: Card Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/card__controller__card_controller__CardOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /exam-sessions:
    get:
      tags:
      - Exams
      summary: List Exam History
      operationId: list_exam_history_exam_sessions_get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/ExamSessionHistoryOut'
                type: array
                title: Response List Exam History Exam Sessions Get
    post:
      tags:
      - Exams
      summary: Start Exam Session
      operationId: start_exam_session_exam_sessions_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartExamSessionIn'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /exam-sessions/{session_id}:
    get:
      tags:
      - Exams
      summary: Get Exam Session
      operationId: get_exam_session_exam_sessions__session_id__get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamSessionOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /exam-sessions/{session_id}/answers/{card_id}:
    put:
      tags:
      - Exams
      summary: Save Exam Answer
      operationId: save_exam_answer_exam_sessions__session_id__answers__card_id__put
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      - name: card_id
        in: path
        required: true
        schema:
          type: string
          title: Card Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SaveAnswerIn'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SaveAnswerOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /exam-sessions/{session_id}/submit:
    post:
      tags:
      - Exams
      summary: Submit Exam Session
      operationId: submit_exam_session_exam_sessions__session_id__submit_post
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResultOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /exam-sessions/{session_id}/result:
    get:
      tags:
      - Exams
      summary: Get Exam Result
      operationId: get_exam_result_exam_sessions__session_id__result_get
      parameters:
      - name: session_id
        in: path
        required: true
        schema:
          type: string
          title: Session Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExamResultOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CardContentOut:
      properties:
        text:
          type: string
          title: Text
        images:
          items:
            $ref: '#/components/schemas/CardImageOut'
          type: array
          title: Images
      type: object
      required:
      - text
      - images
      title: CardContentOut
    CardImageOut:
      properties:
        image_id:
          type: string
          title: Image Id
        storage_key:
          type: string
          title: Storage Key
        alt_text:
          anyOf:
          - type: string
          - type: 'null'
          title: Alt Text
      type: object
      required:
      - image_id
      - storage_key
      title: CardImageOut
...
    ExamQuestionOut:
      properties:
        question_number:
          type: integer
          title: Question Number
        card_id:
          type: string
          title: Card Id
        question_text:
          type: string
          title: Question Text
        question_images:
          items:
            $ref: '#/components/schemas/CardImageOut'
          type: array
          title: Question Images
        student_answer:
          type: string
          title: Student Answer
        answered_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Answered At
      type: object
      required:
      - question_number
      - card_id
      - question_text
      - question_images
      - student_answer
      title: ExamQuestionOut
    ExamQuestionResultOut:
      properties:
        question_number:
          type: integer
          title: Question Number
        card_id:
          type: string
          title: Card Id
        question_text:
          type: string
          title: Question Text
        reference_short_answer:
          items:
            type: string
          type: array
          title: Reference Short Answer
        student_answer:
          type: string
          title: Student Answer
        score:
          type: number
          title: Score
        is_correct:
          type: boolean
          title: Is Correct
        feedback:
          type: string
          title: Feedback
        errors:
          items:
            type: string
          type: array
          title: Errors
      type: object
      required:
      - question_number
      - card_id
      - question_text
      - reference_short_answer
      - student_answer
      - score
      - is_correct
      - feedback
      - errors
      title: ExamQuestionResultOut
    ExamResultOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        time_limit_minutes:
          type: integer
          title: Time Limit Minutes
        time_over:
          type: boolean
          title: Time Over
        total_score:
          type: number
          title: Total Score
        max_score:
          type: number
          title: Max Score
        passed:
          type: boolean
          title: Passed
        pass_score_threshold:
          type: number
          title: Pass Score Threshold
        questions:
          items:
            $ref: '#/components/schemas/ExamQuestionResultOut'
          type: array
          title: Questions
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - time_limit_minutes
      - time_over
      - total_score
      - max_score
      - passed
      - pass_score_threshold
      - questions
      title: ExamResultOut
    ExamSessionHistoryOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        total_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Total Score
        max_score:
          anyOf:
          - type: number
          - type: 'null'
          title: Max Score
        passed:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Passed
        time_over:
          type: boolean
          title: Time Over
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - time_over
      title: ExamSessionHistoryOut
    ExamSessionOut:
      properties:
        session_id:
          type: string
          title: Session Id
        sheet_number:
          type: integer
          title: Sheet Number
        status:
          type: string
          title: Status
        started_at:
          type: string
          title: Started At
        submitted_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Submitted At
        deadline_at:
          type: string
          title: Deadline At
        time_limit_minutes:
          type: integer
          title: Time Limit Minutes
        time_remaining_seconds:
          type: integer
          title: Time Remaining Seconds
        time_over:
          type: boolean
          title: Time Over
        question_count:
          type: integer
          title: Question Count
        questions:
          items:
            $ref: '#/components/schemas/ExamQuestionOut'
          type: array
          title: Questions
      type: object
      required:
      - session_id
      - sheet_number
      - status
      - started_at
      - deadline_at
      - time_limit_minutes
      - time_remaining_seconds
      - time_over
      - question_count
      - questions
      title: ExamSessionOut
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    SaveAnswerIn:
      properties:
        student_answer:
          type: string
          title: Student Answer
          default: ''
      type: object
      title: SaveAnswerIn
    SaveAnswerOut:
      properties:
        session_id:
          type: string
          title: Session Id
        card_id:
          type: string
          title: Card Id
        question_number:
          type: integer
          title: Question Number
        student_answer:
          type: string
          title: Student Answer
        answered_at:
          anyOf:
          - type: string
          - type: 'null'
          title: Answered At
      type: object
      required:
      - session_id
      - card_id
      - question_number
      - student_answer
      title: SaveAnswerOut
...
    StartExamSessionIn:
      properties:
        sheet_number:
          type: integer
          maximum: 15.0
          minimum: 1.0
          title: Sheet Number
        time_limit_minutes:
          anyOf:
          - type: integer
            maximum: 240.0
            minimum: 1.0
          - type: 'null'
          title: Time Limit Minutes
      type: object
      required:
      - sheet_number
      title: StartExamSessionIn
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    card__controller__card_controller__CardOut:
      properties:
        card_id:
          type: string
          title: Card Id
        front:
          $ref: '#/components/schemas/CardContentOut'
        answer:
          $ref: '#/components/schemas/CardContentOut'
        short_answer:
          items:
            type: string
          type: array
          title: Short Answer
        tags:
          items:
            type: string
          type: array
          title: Tags
        exam_sheets:
          items:
            type: integer
          type: array
          title: Exam Sheets
      type: object
      required:
      - card_id
      - front
      - answer
      - short_answer
      - tags
      - exam_sheets
      title: CardOut
...